Add kfont context (part 1)

+++ src/libkfont/psffontop.h
@@ -20,19 +20,19 @@ struct unicode_list {
 	struct unicode_seq *seq;
 };
 
-extern int readpsffont(FILE *fontf, char **allbufp, int *allszp,
-                       char **fontbufp, int *fontszp,
-                       int *fontwidthp, int *fontlenp, int fontpos0,
+extern int readpsffont(FILE *fontf, char **allbufp, size_t *allszp,
+                       char **fontbufp, size_t *fontszp,
+                       size_t *fontwidthp, size_t *fontlenp, size_t fontpos0,
                        struct unicode_list **uclistheadsp);
 
 extern int writepsffont(FILE *ofil, char *fontbuf,
-                        int width, int height, size_t fontlen, int psftype,
+                        size_t width, size_t height, size_t fontlen, int psftype,
                         struct unicode_list *uclistheads);
 
 #define WPSFH_HASTAB 1
 #define WPSFH_HASSEQ 2
 extern int writepsffontheader(FILE *ofil,
-                              int width, int height, int fontlen,
+                              size_t width, size_t height, size_t fontlen,
                               int *psftype, int flags);
 
 extern int appendunicode(FILE *fp, unsigned int uc, int utf8);
